experiment with machine-word type

+++ src/Swarm/Game/CESK.hs
@@ -85,6 +85,7 @@ import Control.Lens.Combinators (pattern Empty)
 import Data.Aeson (FromJSON, ToJSON)
 import Data.IntMap.Strict (IntMap)
 import Data.IntMap.Strict qualified as IM
+import Data.Int (Int64)
 import GHC.Generics (Generic)
 import Prettyprinter (Doc, Pretty (..), encloseSep, hsep, (<+>))
 import Swarm.Game.Entity (Count, Entity)
@@ -101,12 +102,12 @@ import Swarm.Language.Value as V
 
 -- | A newtype representing a count of ticks (typically since the
 --   start of a game).
-newtype TickNumber = TickNumber {getTickNumber :: Integer}
+newtype TickNumber = TickNumber {getTickNumber :: Int64}
   deriving (Eq, Ord, Show, Read, Generic, FromJSON, ToJSON)
 
 -- | Add an offset to a 'TickNumber'.
-addTicks :: Integer -> TickNumber -> TickNumber
-addTicks i (TickNumber n) = TickNumber $ n + i
+addTicks :: Int -> TickNumber -> TickNumber
+addTicks i (TickNumber n) = TickNumber $ n + fromIntegral i
 
 instance Pretty TickNumber where
   pretty (TickNumber i) = pretty i
